Extract some of the drop logic into super classes to make it easier to read what is going on
